Add createStore.js and botReducer.js
